pyproject.toml
[project]
name = "legion-koi"
version = "0.1.0"
description = "Legion's sovereign KOI-net node — knowledge federation for the plugin ecosystem"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "koi-net~=1.2",
    "rid-lib>=3.2.7",
    "fastapi>=0.115.12",
    "pydantic>=2.10.6",
    "uvicorn>=0.34.2",
    "watchdog>=4.0.0",
    "structlog>=25.4.0",
    "psycopg[binary]>=3.2",
    "mcp>=1.0",
    "httpx>=0.27",
    "redis>=5.0",
    "instructor>=1.0",
    "pyyaml>=6.0",
    "rapidfuzz>=3.0",
    "jinja2>=3.1",
]

[project.optional-dependencies]
rerank = ["sentence-transformers>=3.0"]

[tool.hatch.build.targets.wheel]
packages = ["src/legion_koi"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
